More small portability fixes from V2 branch.
[git-p4: depot-paths = "//depot/openwatcom/": change = 33698]

+++ bld/watcom/h/wressetr.h
@@ -44,8 +44,8 @@ struct WResRoutines {                                       /* defaults */
     /* I/O routines */
     WResFileID (*   open) (const char *, int, ...);         /* open */
     int (*          close) (WResFileID);                    /* close */
-    int (*          write) (WResFileID, const void *, size_t); /* write */
-    int (*          read) (WResFileID, void *, size_t);     /* read */
+    ssize_t (*      write) (WResFileID, const void *, size_t); /* write */
+    ssize_t (*      read) (WResFileID, void *, size_t);     /* read */
     off_t (*        seek) (WResFileID, off_t, int );        /* lseek */
     off_t (*        tell) (WResFileID);                     /* tell */
     /* memory routines */
